mux: move old API into a Compat module
And update existing users to use the compat module. We will shift them
over one by one.
