+++ network-mux/test/Test/Mux.hs
@@ -58,7 +58,7 @@ import           System.Process (createPipe)
 
 import           Test.Mux.ReqResp
 
-import           Network.Mux
+import           Network.Mux.Compat
 import           Network.Mux.Codec
 import           Network.Mux.Channel
 import           Network.Mux.Types (MuxSDU(..), MuxSDUHeader(..), RemoteClockModel(..))
