+++ ouroboros-network-framework/src/Ouroboros/Network/Socket.hs
@@ -79,7 +79,7 @@ import qualified Network.Socket as Socket
 
 import           Control.Tracer
 
-import qualified Network.Mux as Mx
+import qualified Network.Mux.Compat as Mx
 import Network.Mux.DeltaQ.TraceTransformer
 import           Network.Mux.Types (MuxBearer)
